chore: Add pre-commit hooks and custom hooks for code formatting and quality checks
This commit introduces a .pre-commit-config.yaml file, which includes configurations for pre-commit hooks responsible for code formatting and quality checks. The pre-commit hooks are sourced from two external repositories: pre-commit/pre-commit-hooks and psf/black.

Hooks Added:
- check-yaml: Verifies YAML files for syntax errors.
- end-of-file-fixer: Ensures that all files end with a newline character.
- trailing-whitespace: Detects and removes trailing whitespaces from the changes being committed.
- check-added-large-files: Prevents the addition of large files to the repository.
- debug-statements: Checks for and removes debug statements from Python files.
- black: Applies the black code formatter to ensure consistent Python code style.

This commit also introduces two custom hooks, './hooks/commit-msg' and './hooks/pre-commit'. Its usage is documented in the ./README.md.

The 'pre-commit' hook script detects whether the email set in the repository is correct or not. It also implements a script to detect if there are any trailing whitespaces in the commit.

Similarly, the 'commit-msg' hook checks if the commit message contains tags for commit types such as “feat”, “fix”, “refactor” etc.