feat: Add code and study docs for intermediate Python topics
This feature update introduces the "./2_intermediate_python" folder, which contains all the code for the 2nd-day session, focusing on the topics:

- args and kwargs
- filter, map, reduce
- ternary operators
- collections
- comprehensions (list comprehension and set comprehension)
- exceptions
- file I/O (reading and writing JSON and CSV files)
- OOP principles (encapsulation and inheritance)

Additionally, the ./nbs/ folder contains Jupyter development notebooks for reference and experimentation.

The ./1_basics/main.py/ file is moved from the root directory for a topic-wise project folder structure.

The ./study_docs/ folder contains PDF of study materials associated and curated from contents at fuse-classroom.

The ./README.md contains handy scripts that are necessary for easy setup and execution of code.