feat: Learn Lambda, S3 Buckets, and RDS Usage
In this commit, we explored the practical usage of AWS Lambda, S3 Buckets, and RDS in the following files:

- ./7_aws_lambda_rds/1_bucket_lambda.py: We demonstrated how to create a Lambda function that interacts with S3 buckets. We covered loading data from an external API (specifically, the stock_data API from https://rapidapi.com/suneetk92/api/latest-stock-price/) and storing it in the apprentice-training-ml-dev-utsav-raw-data S3 bucket. The cleaned version of the data was stored in the apprentice-training-ml-dev-utsav-cleaned-data S3 bucket.

- ./7_aws_lambda_rds/2_rds.py: We delved into using AWS RDS (Relational Database Service) to establish a connection to a PostgreSQL database. We learned how to define and create a table (etl_utsav_stock_data) in the RDS instance. Additionally, we demonstrated the process of inserting cleaned data from the S3 bucket into the RDS database table.

Furthermore, we utilized the https://rapidapi.com/suneetk92/api/latest-stock-price/ API to fetch stock data, which was then stored in the ./data/aws_data/raw_data directory. The cleaned version of the data was stored in the ./data/aws_data/cleaned_data directory.

These changes provide a comprehensive understanding of integrating Lambda, S3, and RDS for data processing and storage.